(publishing) Remove bintray plugin and set bintray credentials using maven-publish instead
The bintray plugin seems to be incompatible with the new Android maven-publish plugin. Setting up the credentials and URL manually isn't difficult, so I've just removed the plugin. This change doesn't affect the end required environment variables - it just means that the bintray plugin no longer needs to be added to the project dependnencies.